refactor(onboarding): delete bespoke PHP component/field tree
The Components/ tree (ComponentFactory, Page/Section/Field/Fields/*) was a
parallel copy of the settings element system that steps no longer use — they
harvest the canonical schema instead. Retarget SettingsElementTest to in-file
doubles so it no longer depends on the deleted concrete elements.
