--- includes/Admin/OnboardingSetup/Components/FieldGroup.php
@@ -1,51 +1,0 @@
-<?php
-
-namespace WeDevs\Dokan\Admin\OnboardingSetup\Components;
-
-use WeDevs\Dokan\Abstracts\SettingsElement;
-
-/**
- * Field Group Class.
- */
-class FieldGroup extends SettingsElement {
-
-	/**
-	 * Settings Element type.
-	 *
-	 * @var string $type Settings Element type.
-	 */
-	protected $type = 'fieldgroup';
-
-	/**
-	 * Data validation.
-	 *
-	 * @param mixed $data Data for validation.
-	 *
-	 * @return bool
-	 */
-	public function data_validation( $data ): bool {
-		return is_array( $data );
-	}
-
-	/**
-	 * Sanitize data for storage.
-	 *
-	 * @param mixed $data Data for sanitization.
-	 *
-	 * @return array|string
-	 */
-	public function sanitize_element( $data ) {
-		return wp_unslash( $data );
-	}
-
-	/**
-	 * Escape data for display.
-	 *
-	 * @param array $data Data for display.
-	 *
-	 * @return array
-	 */
-	public function escape_element( $data ): array {
-		return $data;
-	}
-}
