--- includes/Admin/OnboardingSetup/Components/Fields/RadioBox.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace WeDevs\Dokan\Admin\OnboardingSetup\Components\Fields;
-
-/**
- * RadioButton Field.
- *
- * Custom field that provides radio options as button.
- */
-class RadioBox extends Radio {
-
-    /**
-     * Input Type.
-     *
-     * @var string $input_type Input Type.
-     */
-    protected $input_type = 'radio_box';
-}
